requirements.txt
numpy ~= 1.14.3
scipy ~= 1.0.1
future ~= 0.16.0
matplotlib ~= 2.2.2
pillow >= 6.2.0
opencv-python ~= 3.4.0
scikit-image ~= 0.14.0
pyaml
scikit-learn
torch ~= 1.4.0
torchvision ~= 0.5.0